Guard

+++ src/brand/brand.controller.ts
@@ -1,34 +1,56 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Req, ForbiddenException } from '@nestjs/common';
 import { BrandService } from './brand.service';
 import { CreateBrandDto } from './dto/create-brand.dto';
 import { UpdateBrandDto } from './dto/update-brand.dto';
+import { ApiBearerAuth } from '@nestjs/swagger';
+import { AuthGuard } from 'src/auth/jwt.guard';
+import { Request, request } from 'express';
+import { UserRole } from '@prisma/client';
 
 @Controller('brand')
 export class BrandController {
-  constructor(private readonly brandService: BrandService) {}
+  constructor(private readonly brandService: BrandService) { }
 
+
+  @ApiBearerAuth()
+  @UseGuards(AuthGuard)
   @Post()
-  create(@Body() createBrandDto: CreateBrandDto) {
-    return this.brandService.create(createBrandDto);
+  create(@Req() req: Request, @Body() createBrandDto: CreateBrandDto) {
+    if (req.user?.role == UserRole.ADMIN) {
+      return this.brandService.create(createBrandDto);
+    } else {
+      throw new ForbiddenException("Access denied")
+    }
   }
 
   @Get()
   findAll() {
     return this.brandService.findAll();
   }
-
   @Get(':id')
   findOne(@Param('id') id: string) {
     return this.brandService.findOne(+id);
   }
 
+  @ApiBearerAuth()
+  @UseGuards(AuthGuard)
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateBrandDto: UpdateBrandDto) {
-    return this.brandService.update(+id, updateBrandDto);
+  update(@Req() req: Request, @Param('id') id: string, @Body() updateBrandDto: UpdateBrandDto) {
+    if (req.user?.role == UserRole.ADMIN || req.user?.role == UserRole.SUPER_ADMIN) {
+      return this.brandService.update(+id, updateBrandDto);
+    } else {
+      throw new ForbiddenException("Access denied")
+    }
   }
 
+  @ApiBearerAuth()
+  @UseGuards(AuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.brandService.remove(+id);
+  remove(@Req() req: Request, @Param('id') id: string) {
+    if (req.user?.role == UserRole.ADMIN || req.user?.role == UserRole.SUPER_ADMIN) {
+      return this.brandService.remove(+id);
+    } else {
+      throw new ForbiddenException("Access denied")
+    }
   }
 }
