+++ src/general-info/general-info.controller.ts
@@ -1,9 +1,11 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Req, ForbiddenException } from '@nestjs/common';
 import { GeneralInfoService } from './general-info.service';
 import { CreateGeneralInfoDto } from './dto/create-general-info.dto';
 import { UpdateGeneralInfoDto } from './dto/update-general-info.dto';
 import { ApiBearerAuth } from '@nestjs/swagger';
 import { AuthGuard } from 'src/auth/jwt.guard';
+import { Request } from 'express';
+import { UserRole } from '@prisma/client';
 
 @Controller('general-info')
 export class GeneralInfoController {
@@ -12,8 +14,12 @@ export class GeneralInfoController {
   @ApiBearerAuth()
   @UseGuards(AuthGuard)
   @Post()
-  create(@Body() createGeneralInfoDto: CreateGeneralInfoDto) {
-    return this.generalInfoService.create(createGeneralInfoDto);
+  create(@Req() req: Request, @Body() createGeneralInfoDto: CreateGeneralInfoDto) {
+    if (req.user?.role == UserRole.ADMIN) {
+      return this.generalInfoService.create(createGeneralInfoDto);
+    } else {
+      throw new ForbiddenException("Access denied")
+    }
   }
 
   @Get()
@@ -29,14 +35,22 @@ export class GeneralInfoController {
   @ApiBearerAuth()
   @UseGuards(AuthGuard)
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateGeneralInfoDto: UpdateGeneralInfoDto) {
-    return this.generalInfoService.update(+id, updateGeneralInfoDto);
+  update(@Req() req: Request, @Param('id') id: string, @Body() updateGeneralInfoDto: UpdateGeneralInfoDto) {
+    if (req.user?.role == UserRole.ADMIN || req.user?.role == UserRole.SUPER_ADMIN) {
+      return this.generalInfoService.update(+id, updateGeneralInfoDto);
+    } else {
+      throw new ForbiddenException("Access denied")
+    }
   }
 
   @ApiBearerAuth()
   @UseGuards(AuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.generalInfoService.remove(+id);
+  remove(@Req() req: Request, @Param('id') id: string) {
+    if (req.user?.role == UserRole.ADMIN || req.user?.role == UserRole.SUPER_ADMIN) {
+      return this.generalInfoService.remove(+id);
+    } else {
+      throw new ForbiddenException("Access denied")
+    }
   }
 }
